solved problme of migrations and resolved #66

+++ database/migrations/2024_10_22_112852_create_production_component_consumptions_table.php
@@ -4,25 +4,26 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::create('component_consumptions', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('production_batch_id')->constrained()->onDelete('cascade');
-            $table->morphs('component');
-            $table->foreignId('inventory_batch_id')->constrained();
-            $table->decimal('quantity', 10, 3);
-            $table->decimal('price_per_unit', 10, 2);
-            $table->foreignId('unit_id')->constrained();
-            $table->decimal('waste_quantity', 10, 3)->default(0);
-            $table->text('notes')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('component_consumptions')) {
+            Schema::create('component_consumptions', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('production_batch_id')->constrained()->onDelete('cascade');
+                $table->morphs('component');
+                $table->foreignId('inventory_batch_id')->constrained();
+                $table->decimal('quantity', 10, 3);
+                $table->decimal('price_per_unit', 10, 2);
+                $table->foreignId('unit_id')->constrained();
+                $table->decimal('waste_quantity', 10, 3)->default(0);
+                $table->text('notes')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
