+++ database/migrations/2024_10_22_115441_create_component_reservations_table.php
@@ -4,22 +4,23 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::create('component_reservations', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('production_batch_id')->constrained()->onDelete('cascade');
-            $table->morphs('component');
-            $table->decimal('quantity', 10, 3);
-            $table->foreignId('unit_id')->constrained();
-            $table->timestamp('expires_at');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('component_reservations')) {
+            Schema::create('component_reservations', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('production_batch_id')->constrained()->onDelete('cascade');
+                $table->morphs('component');
+                $table->decimal('quantity', 10, 3);
+                $table->foreignId('unit_id')->constrained();
+                $table->timestamp('expires_at');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
